ci: gate on fmt and clippy, and make the workspace pass them
Nothing checked formatting, so the workspace had drifted out of rustfmt shape
— running `cargo fmt --all` while adding the climate module reformatted 24
files across all four crates. Rather than revert that, this adds the gate that
would have prevented the drift and fixes the lints it turns up.

The lint job runs first and installs no tshark: a formatting slip should fail
in seconds, not after an apt-get and a four-minute test cycle. Clippy runs
with --all-targets and -D warnings, so the examples are held to the library's
standard — they are the series' evidence artifacts, not scratch code.

Fourteen warnings cleared:

- needless_borrow x4 in the failure and feeder examples (clippy --fix).
- needless_range_loop x2, same pattern in both: `for ring in 0..wheel.planes`
  indexing `phases`. plane_phases() returns exactly one entry per plane, by
  construction and by its own doc, so iterating the vec is equivalent.
- unnecessary_cast x3. Removing the casts left the array literals with an
  unconstrained numeric type (E0689), so the element type is pinned on the
  first literal instead of re-cast at every use.
- type_complexity x4. Three are one nameable type — PairSamples, the
  time-ascending sample list behind ChannelTrace. The fourth is netif's test
  bench(), whose tuple carries four `impl EventSinkReader<...>` members that
  cannot be named in a stable type alias; allowed with that reason recorded.
- doc list continuation x2, in coverage.rs and records.rs.

Verified before committing: fmt --check clean, clippy -D warnings clean,
cargo test --workspace green at 231 tests, and failure_timeline still prints
its published evidence — the 394 ms detour floor, 430/460 ms detour round
trips, the 37,294 km frozen plane link.

+++ crates/orbits/examples/compute_placement.rs
@@ -34,11 +34,11 @@ fn main() {
 
     println!("Anchor candidates:");
     for (label, sep_deg) in [
-        ("MEO 20,000 km, overhead", 0.0),
+        ("MEO 20,000 km, overhead", 0.0_f64),
         ("MEO 20,000 km, 30° away", 30.0),
         ("MEO 20,000 km, 60° away", 60.0),
     ] {
-        let d = shell_distance(&planet, ACCESS_ALT, MEO_ALT, (sep_deg as f64).to_radians());
+        let d = shell_distance(&planet, ACCESS_ALT, MEO_ALT, sep_deg.to_radians());
         println!(
             "  {label:<28} {:>8.0} km  {:>6.1} ms one way from access",
             d / 1e3,
